Extract ImmediateFillModel into new base FillModel class
The new FillModel class will serve as a base for all fill models.
This provides benefits in backwards compatibility by providing a common
place for 'shim' code. The ImmediateFillModel class remains but is now
empty as it's entire implementation has been moved to the new FillModel
class.
