+++ Common/QuantConnect.csproj
@@ -214,6 +214,7 @@
     <Compile Include="Interfaces\IRegressionAlgorithmDefinition.cs" />
     <Compile Include="Interfaces\ITimeInForceHandler.cs" />
     <Compile Include="Orders\Fees\BitfinexFeeModel.cs" />
+    <Compile Include="Orders\Fills\FillModel.cs" />
     <Compile Include="Orders\Fills\LatestPriceFillModel.cs" />
     <Compile Include="Orders\BitfinexOrderProperties.cs" />
     <Compile Include="Orders\GDAXOrderProperties.cs" />
